refactor(client): split conversations_provider.dart into 3 files
conversations_ws_handlers.dart (124 LoC): WS event handlers
(onNewMessage, onMessageEdited, updateDecryptedPreview,
updateEncryption, addGroupMember) extracted as mixin via part/part-of.

conversations_http_actions.dart (326 LoC): REST action methods
(getOrCreateDm, leaveConversation, toggleMute, setMuted, setPinned,
leaveGroup, createGroup) extracted as second mixin.

conversations_provider.dart reduced from 657 to 256 LoC. Public API
unchanged -- all 30+ consumer call sites import same file, no
widget changes required. 1140 tests pass.
